refactor: move command types to the command files
rel #2154
