+++ packages/pnpm/src/cmd/import.ts
@@ -1,14 +1,20 @@
+import { types as allTypes } from '@pnpm/config'
 import { WANTED_LOCKFILE } from '@pnpm/constants'
 import PnpmError from '@pnpm/error'
 import { readImporterManifestOnly } from '@pnpm/read-importer-manifest'
 import rimraf = require('@zkochan/rimraf')
 import loadJsonFile = require('load-json-file')
 import path = require('path')
+import R = require('ramda')
 import renderHelp = require('render-help')
 import { install } from 'supi'
 import createStoreController from '../createStoreController'
 import { PnpmOptions } from '../types'
 import { docsUrl } from './help'
+
+export function types () {
+  return {}
+}
 
 export function help () {
   return renderHelp({
