+++ packages/pnpm/src/cmd/index.ts
@@ -25,7 +25,7 @@ import * as unlink from './unlink'
 import * as update from './update'
 import * as why from './why'
 
-const commands: Array<{ commandNames: string[], handler: Function, help: () => string}> = [
+const commands: Array<{ commandNames: string[], handler: Function, help: () => string, types: () => Object}> = [
   add,
   audit,
   importCmd,
@@ -55,15 +55,32 @@ const commands: Array<{ commandNames: string[], handler: Function, help: () => s
 
 const handlerByCommandName: Record<string, Function> = {}
 const helpByCommandName: Record<string, () => string> = {}
+const typesByCommandName: Record<string, () => Object> = {}
+const aliasToFullName: Map<string, string> = new Map()
 
-for (const { commandNames, handler, help } of commands) {
+for (const { commandNames, handler, help, types } of commands) {
   if (commandNames.length === 0) throw new Error('One of the commands doesn\'t have command names')
   for (const commandName of commandNames) {
     handlerByCommandName[commandName] = handler
     helpByCommandName[commandName] = help
+    typesByCommandName[commandName] = types
+  }
+  if (commandNames.length > 1) {
+    const fullName = commandNames[0]
+    for (let i = 1; i < commandNames.length; i++) {
+      aliasToFullName.set(commandNames[i], fullName)
+    }
   }
 }
 
 handlerByCommandName.help = createHelp(helpByCommandName)
 
 export default handlerByCommandName
+
+export function getTypes (commandName: string) {
+  return typesByCommandName[commandName]?.() || {}
+}
+
+export function getCommandFullName (commandName: string) {
+  return aliasToFullName.get(commandName) || commandName
+}
