+++ packages/pnpm/src/cmd/rebuild.ts
@@ -1,4 +1,6 @@
+import { types as allTypes } from '@pnpm/config'
 import { oneLine } from 'common-tags'
+import R = require('ramda')
 import renderHelp = require('render-help')
 import {
   rebuild,
@@ -8,6 +10,12 @@ import createStoreController from '../createStoreController'
 import { readImporterManifestOnly } from '../readImporterManifest'
 import { PnpmOptions } from '../types'
 import { docsUrl, FILTERING, UNIVERSAL_OPTIONS } from './help'
+
+export function types () {
+  return R.pick([
+    'recursive',
+  ], allTypes)
+}
 
 export const commandNames = ['rebuild', 'rb']
