+++ packages/pnpm/src/cmd/stop.ts
@@ -2,6 +2,10 @@ import renderHelp = require('render-help')
 import { docsUrl } from './help'
 import { handler as run } from './run'
 
+export function types () {
+  return {}
+}
+
 export const commandNames = ['stop']
 
 export function help () {
