+++ packages/pnpm/src/cmd/test.ts
@@ -1,7 +1,15 @@
+import { types as allTypes } from '@pnpm/config'
 import { oneLine } from 'common-tags'
+import R = require('ramda')
 import renderHelp = require('render-help')
 import { docsUrl, FILTERING } from './help'
 import { handler as run } from './run'
+
+export function types () {
+  return R.pick([
+    'recursive',
+  ], allTypes)
+}
 
 export const commandNames = ['test', 't', 'tst']
