+++ packages/pnpm/src/cmd/why.ts
@@ -1,7 +1,24 @@
+import { types as allTypes } from '@pnpm/config'
 import { oneLine, stripIndent } from 'common-tags'
+import R = require('ramda')
 import renderHelp = require('render-help')
 import { docsUrl, FILTERING, OPTIONS, UNIVERSAL_OPTIONS } from './help'
 import { handler as list } from './list'
+
+export function types () {
+  return R.pick([
+    'dev',
+    'global-dir',
+    'global',
+    'json',
+    'long',
+    'only',
+    'optional',
+    'parseable',
+    'production',
+    'recursive',
+  ], allTypes)
+}
 
 export const commandNames = ['why']
